feat: update deploy script to do some git-fu

+++ buildDeployment.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ `git status --porcelain` ]]; then
+  echo "Other git changes detected, please isolate deployments to single commits."
+  exit 1;
+fi
+
 rm -rf dist/
 
 rm -rf docs/
@@ -8,3 +13,10 @@ mkdir docs
 ng build --prod
 
 mv dist/* docs
+
+RELEASE_MSG=`date "+%Y-%m-%d (%s)"`
+
+git add .
+git commit -am "release: $RELEASE_MSG"
+
+echo "Deployment built and committed to local branch, push to remote to deploy to GitHub pages!"
